new get chunk
