+++ pumpkin-world/Cargo.toml
@@ -47,9 +47,9 @@ lru = "0.16.0"
 tokio-util = { version = "0.7.16", features = ["rt"] }
 rand = "0.9.2"
 num_cpus = "1.17.0"
-nohash-hasher = "0.2.0"
 slotmap = "1.0.7"
 intrusive-collections = "0.9.7"
+rustc-hash = "2.1.1"
 
 [dev-dependencies]
 criterion = { version = "0.7", default-features = false, features = ["html_reports", "async_tokio"] }
