+++ pumpkin-world/src/chunk_system.rs
@@ -1,0 +1,1288 @@
+use crate::block::RawBlockState;
+use crate::chunk::ChunkHeightmapType;
+use crate::chunk::io::FileIO;
+use crate::chunk::io::LoadedData::Loaded;
+use crate::dimension::Dimension;
+
+use crate::generation::height_limit::HeightLimitView;
+
+use crate::generation::proto_chunk::{GenerationCache, StagedChunk, TerrainCache};
+use crate::generation::settings::{GenerationSettings, gen_settings_from_dimension};
+use crate::level::{Level, SyncChunk};
+use crate::world::{BlockAccessor, BlockRegistryExt};
+use crate::{GlobalRandomConfig, ProtoChunk, ProtoNoiseRouters};
+use async_trait::async_trait;
+use crossbeam::channel::{Receiver, SendError, Sender};
+use dashmap::DashMap;
+use itertools::Itertools;
+use log::debug;
+use num_traits::abs;
+use pumpkin_data::biome::Biome;
+
+use pumpkin_data::{Block, BlockState};
+use pumpkin_util::HeightMap;
+use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::vector2::Vector2;
+use pumpkin_util::math::vector3::Vector3;
+
+use std::cmp::{Ordering, PartialEq, max, min};
+use std::collections::hash_map::Entry;
+use std::collections::{BinaryHeap, HashMap, HashSet, VecDeque};
+use std::mem::swap;
+use std::sync::{Arc, Condvar, Mutex};
+
+use crate::chunk_system::StagedChunkEnum::{Biomes, Empty, Features, Noise, Surface};
+use rustc_hash::FxHashMap;
+use std::hash::BuildHasherDefault;
+use std::sync::atomic::Ordering::Relaxed;
+use std::thread::JoinHandle;
+use std::time::{Duration, Instant};
+use std::{ptr, thread};
+use tokio::runtime::{Builder, Runtime};
+use tokio::sync::{RwLock, oneshot};
+use tokio::task;
+use tokio_util::task::TaskTracker;
+
+type HashMapType<K, V> = FxHashMap<K, V>;
+type ChunkPos = Vector2<i32>;
+type ChunkLevel = HashMapType<ChunkPos, i8>;
+
+struct HeapNode(i8, ChunkPos);
+impl PartialEq for HeapNode {
+    fn eq(&self, other: &Self) -> bool {
+        self.0 == other.0
+    }
+}
+impl Eq for HeapNode {}
+impl PartialOrd for HeapNode {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+impl Ord for HeapNode {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.0.cmp(&other.0).reverse()
+    }
+}
+impl From<(ChunkPos, i8)> for HeapNode {
+    fn from(value: (ChunkPos, i8)) -> Self {
+        Self {
+            0: value.1,
+            1: value.0,
+        }
+    }
+}
+impl Into<(ChunkPos, i8)> for HeapNode {
+    fn into(self) -> (ChunkPos, i8) {
+        (self.1, self.0)
+    }
+}
+
+pub struct ChunkLoading {
+    pub is_dirty: bool,
+    pub pos_level: ChunkLevel,
+    pub ticket: HashMapType<ChunkPos, Vec<i8>>, // TODO lifetime & id
+    pub sender: Arc<LevelChannel>,
+}
+
+impl ChunkLoading {
+    pub const FULL_CHUNK_LEVEL: i8 = 33;
+    pub const MAX_LEVEL: i8 = 46; // level 46 will be unloaded.
+    pub fn dump_level_debug(map: &ChunkLevel, sx: i32, tx: i32, sy: i32, ty: i32) {
+        let mut header = "X/Y".to_string();
+        for y in sy..=ty {
+            header.push_str(&format!("{y:4}"));
+        }
+
+        let grid: String = (sx..=tx)
+            .map(|x| {
+                let mut row = format!("{x:3}");
+                row.push_str(
+                    &(sy..=ty)
+                        .map(|y| {
+                            format!(
+                                "{:4}",
+                                map.get(&ChunkPos::new(x, y))
+                                    .unwrap_or(&ChunkLoading::MAX_LEVEL)
+                            )
+                        })
+                        .collect::<String>(),
+                );
+                row
+            })
+            .collect::<Vec<_>>()
+            .join("\n");
+
+        debug!("\nloading level:\n{header}\n{grid}");
+    }
+
+    pub const fn get_level_from_view_distance(view_distance: u8) -> i8 {
+        Self::FULL_CHUNK_LEVEL + 1 - (view_distance as i8)
+    }
+
+    pub fn new(sender: Arc<LevelChannel>) -> Self {
+        Self {
+            is_dirty: true,
+            pos_level: ChunkLevel::default(),
+            ticket: HashMapType::default(),
+            sender,
+        }
+    }
+
+    pub fn send_change(&mut self) {
+        if self.is_dirty {
+            self.is_dirty = false;
+            self.sender.set(self.get_cloned_level());
+        }
+    }
+
+    fn run_update(pos_level: &mut ChunkLevel, mut queue: BinaryHeap<HeapNode>) {
+        while !queue.is_empty() {
+            let (pos, level) = unsafe { queue.pop().unwrap_unchecked().into() }; // we have checked. unsafe is safe
+            debug_assert!(level < Self::MAX_LEVEL);
+            if level > *pos_level.get(&pos).unwrap_or(&Self::MAX_LEVEL) {
+                continue;
+            }
+            debug_assert_eq!(level, *pos_level.get(&pos).unwrap_or(&Self::MAX_LEVEL));
+            let spread_level = level + 1;
+            if spread_level >= Self::MAX_LEVEL {
+                continue;
+            }
+            for dx in -1..2 {
+                for dy in -1..2 {
+                    let new_pos = pos.add_raw(dx, dy);
+                    if new_pos != pos {
+                        Self::check_then_push(pos_level, &mut queue, new_pos, spread_level);
+                    }
+                }
+            }
+        }
+    }
+
+    fn check_then_push(
+        pos_level: &mut ChunkLevel,
+        queue: &mut BinaryHeap<HeapNode>,
+        pos: ChunkPos,
+        level: i8,
+    ) {
+        match pos_level.entry(pos) {
+            Entry::Occupied(mut entry) => {
+                let old = entry.get_mut();
+                if *old <= level {
+                    return;
+                }
+                *old = level;
+            }
+            Entry::Vacant(empty) => {
+                empty.insert(level);
+            }
+        }
+        queue.push((pos, level).into());
+    }
+
+    fn run_increase_update(&mut self, pos: ChunkPos, level: i8) {
+        // TODO there will be a faster way
+        debug_assert!(level < Self::MAX_LEVEL);
+        let range = Self::MAX_LEVEL - level - 1;
+        let mut queue = BinaryHeap::new();
+        for dx in -range..=range {
+            for dy in -range..=range {
+                let new_pos = pos.add_raw(dx as i32, dy as i32);
+                let level_from_source = level + abs(dx).max(abs(dy));
+                match self.pos_level.entry(new_pos) {
+                    Entry::Occupied(mut entry) => {
+                        let old = *entry.get();
+                        if old < level_from_source {
+                            queue.push((new_pos, old).into());
+                            continue;
+                        }
+                        debug_assert!(old == level_from_source);
+                        entry.remove();
+                    }
+                    Entry::Vacant(_) => {
+                        panic!("pos {new_pos:?} should contain a level");
+                    }
+                }
+            }
+        }
+        let min_x = pos.x - range as i32 - 1;
+        let max_x = pos.x + range as i32 + 1;
+        let min_y = pos.y - range as i32 - 1;
+        let max_y = pos.y + range as i32 + 1;
+        for y in min_y..max_y {
+            let mut new_pos = ChunkPos::new(max_x, y);
+            let level = self.pos_level.get(&new_pos).unwrap_or(&Self::MAX_LEVEL) + 1;
+            if level < Self::MAX_LEVEL {
+                new_pos.x -= 1;
+                Self::check_then_push(&mut self.pos_level, &mut queue, new_pos, level);
+            }
+        }
+        for x in min_x..max_x {
+            let mut new_pos = ChunkPos::new(x, min_y);
+            let level = self.pos_level.get(&new_pos).unwrap_or(&Self::MAX_LEVEL) + 1;
+            if level < Self::MAX_LEVEL {
+                new_pos.y += 1;
+                Self::check_then_push(&mut self.pos_level, &mut queue, new_pos, level);
+            }
+        }
+        for y in (min_y + 1)..=max_y {
+            let mut new_pos = ChunkPos::new(min_x, y);
+            let level = self.pos_level.get(&new_pos).unwrap_or(&Self::MAX_LEVEL) + 1;
+            if level < Self::MAX_LEVEL {
+                new_pos.x += 1;
+                Self::check_then_push(&mut self.pos_level, &mut queue, new_pos, level);
+            }
+        }
+        for x in (min_x + 1)..=max_x {
+            let mut new_pos = ChunkPos::new(x, max_y);
+            let level = self.pos_level.get(&new_pos).unwrap_or(&Self::MAX_LEVEL) + 1;
+            if level < Self::MAX_LEVEL {
+                new_pos.y -= 1;
+                Self::check_then_push(&mut self.pos_level, &mut queue, new_pos, level);
+            }
+        }
+        for (ticket_pos, levels) in &self.ticket {
+            if abs(ticket_pos.x - pos.x) <= range as i32
+                && abs(ticket_pos.y - pos.y) <= range as i32
+            {
+                Self::check_then_push(
+                    &mut self.pos_level,
+                    &mut queue,
+                    *ticket_pos,
+                    *levels.iter().max().unwrap(),
+                );
+            }
+        }
+        Self::run_update(&mut self.pos_level, queue);
+    }
+    pub fn add_ticket(&mut self, pos: ChunkPos, level: i8) {
+        log::debug!("add ticket at {pos:?} level {level}");
+        debug_assert!(level < Self::MAX_LEVEL);
+        match self.ticket.entry(pos) {
+            Entry::Occupied(mut vec) => {
+                vec.get_mut().push(level);
+            }
+            Entry::Vacant(empty) => {
+                empty.insert(vec![level]);
+            }
+        }
+        match self.pos_level.entry(pos) {
+            Entry::Occupied(mut entry) => {
+                let old = entry.get_mut();
+                if *old < level {
+                    return;
+                }
+                self.is_dirty = true;
+                *old = level;
+            }
+            Entry::Vacant(empty) => {
+                empty.insert(level);
+            }
+        }
+        let mut queue: BinaryHeap<HeapNode> = BinaryHeap::new();
+        queue.push((pos, level).into());
+        Self::run_update(&mut self.pos_level, queue);
+    }
+    pub fn remove_ticket(&mut self, pos: ChunkPos, level: i8) {
+        log::debug!("remove ticket at {pos:?} level {level}");
+        debug_assert!(level < Self::MAX_LEVEL);
+        let Some(vec) = self.ticket.get_mut(&pos) else {
+            log::warn!("No ticket found at {pos:?}");
+            return;
+        };
+        let Some((index, _)) = vec.iter().find_position(|x| **x == level) else {
+            log::warn!("No ticket found at {pos:?}");
+            return;
+        };
+        vec.remove(index);
+        let old_level = *self.pos_level.get(&pos).unwrap_or(&Self::MAX_LEVEL);
+        let source = *vec.iter().max().unwrap_or(&Self::MAX_LEVEL);
+        if vec.is_empty() {
+            self.ticket.remove(&pos);
+        }
+        if level == old_level && source != level {
+            self.is_dirty = true;
+            self.run_increase_update(pos, old_level);
+        }
+    }
+    fn get_cloned_level(&self) -> ChunkLevel {
+        self.pos_level.clone()
+    }
+}
+
+#[test]
+fn test_loading_level() {
+    let mut a = ChunkLoading::new();
+    a.add_ticket(ChunkPos::new(1, 0), 41);
+    println!("add 1,0 41");
+    ChunkLoading::dump_level_debug(&a.pos_level, -20, 20, -20, 20);
+    println!("add 0,-1 41");
+    a.add_ticket(ChunkPos::new(0, -1), 41);
+    ChunkLoading::dump_level_debug(&a.pos_level, -20, 20, -20, 20);
+    println!("remove 1,0 41");
+    a.remove_ticket(ChunkPos::new(1, 0), 41);
+    ChunkLoading::dump_level_debug(&a.pos_level, -20, 20, -20, 20);
+}
+
+#[repr(u8)]
+#[derive(Debug, Copy, Clone, Eq, Ord, PartialEq, PartialOrd)]
+pub enum StagedChunkEnum {
+    None,
+    /// Initial empty chunk, ready for biome population
+    Empty = 1, // EMPTY STRUCTURE_STARTS STRUCTURE_REFERENCES
+    /// Chunk with biomes populated, ready for noise generation
+    Biomes,
+    /// Chunk with terrain noise generated, ready for surface building
+    Noise,
+    /// Chunk with surface built, ready for features and structures
+    Surface, // SURFACE CARVERS
+    /// Chunk with features and structures, ready for finalization
+    Features, // FEATURES INITIALIZE_LIGHT LIGHT SPAWN
+    /// Fully generated chunk
+    Full,
+}
+
+impl From<u8> for StagedChunkEnum {
+    fn from(v: u8) -> Self {
+        match v {
+            1 => Self::Empty,
+            2 => Self::Biomes,
+            3 => Self::Noise,
+            4 => Self::Surface,
+            5 => Self::Features,
+            6 => Self::Full,
+            _ => panic!(),
+        }
+    }
+}
+impl StagedChunkEnum {
+    const MAX_STAGE_ID: u8 = Self::Full as u8;
+    fn level_to_stage(level: i8) -> Self {
+        if level <= 33 {
+            Self::Full
+        } else if level <= 35 {
+            Self::Features
+        } else if level <= 36 {
+            Self::Surface
+        } else if level <= 37 {
+            Self::Biomes
+        } else if level <= 45 {
+            Self::Empty
+        } else {
+            Self::None
+        }
+    }
+    fn get_radius(self) -> i32 {
+        // 这个是不包括自己的范围
+        match self {
+            StagedChunkEnum::Empty => 0,
+            StagedChunkEnum::Biomes => 8,
+            StagedChunkEnum::Noise => 9,
+            StagedChunkEnum::Surface => 9,
+            StagedChunkEnum::Features => 10,
+            StagedChunkEnum::Full => 11,
+            _ => panic!(),
+        }
+    }
+    fn get_dependencies(self) -> &'static [StagedChunkEnum] {
+        match self {
+            StagedChunkEnum::Empty => &[],
+            StagedChunkEnum::Biomes => &[
+                Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty,
+            ],
+            StagedChunkEnum::Noise => &[
+                Biomes, Biomes, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty,
+            ],
+            StagedChunkEnum::Surface => &[
+                Noise, Biomes, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty,
+            ],
+            StagedChunkEnum::Features => &[
+                Surface, Surface, Biomes, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty,
+            ],
+            StagedChunkEnum::Full => &[
+                Features, Features, Surface, Biomes, Empty, Empty, Empty, Empty, Empty, Empty,
+                Empty, Empty,
+            ],
+            _ => panic!(),
+        }
+    }
+}
+
+pub struct LevelChannel {
+    pub value: Mutex<Option<ChunkLevel>>,
+    pub notify: Condvar,
+}
+
+impl Default for LevelChannel {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl LevelChannel {
+    pub fn new() -> Self {
+        Self {
+            value: Mutex::new(None),
+            notify: Condvar::new(),
+        }
+    }
+
+    pub fn set(&self, new_value: ChunkLevel) {
+        *self.value.lock().unwrap() = Some(new_value);
+        self.notify.notify_one();
+    }
+    pub fn get(&self) -> Option<ChunkLevel> {
+        let mut lock = self.value.lock().unwrap();
+        let mut ret = None;
+        swap(&mut ret, &mut *lock);
+        ret
+    }
+    pub fn wait_and_get(&self, level: &Arc<Level>) -> Option<ChunkLevel> {
+        // TODO if this return None, stop thread
+        let mut lock = self.value.lock().unwrap();
+        while lock.is_none() && !level.is_shutting_down.load(Relaxed) {
+            lock = self.notify.wait(lock).unwrap();
+        }
+        let mut ret = None;
+        swap(&mut ret, &mut *lock);
+        ret
+    }
+    pub fn notify(&self) {
+        drop(self.value.lock());
+        self.notify.notify_one();
+    }
+}
+
+pub enum Chunk {
+    Level(SyncChunk),
+    Proto(StagedChunk),
+}
+
+impl Chunk {
+    fn to_stage_id(chunk: &Option<Self>) -> u8 {
+        match chunk {
+            None => 0,
+            Some(Chunk::Proto(data)) => data.stage_id(),
+            Some(Chunk::Level(_)) => 6,
+        }
+    }
+}
+
+struct Cache {
+    x: i32,
+    y: i32,
+    size: i32,
+    pub chunks: Vec<Chunk>,
+    rt: Runtime,
+}
+
+impl HeightLimitView for Cache {
+    fn height(&self) -> u16 {
+        let mid = ((self.size * self.size) >> 1) as usize;
+        match &self.chunks[mid] {
+            Chunk::Proto(chunk) => chunk.proto_chunk().height(),
+            _ => panic!(),
+        }
+    }
+
+    fn bottom_y(&self) -> i8 {
+        let mid = ((self.size * self.size) >> 1) as usize;
+        match &self.chunks[mid] {
+            Chunk::Proto(chunk) => chunk.proto_chunk().bottom_y(),
+            _ => panic!(),
+        }
+    }
+}
+
+#[async_trait]
+impl BlockAccessor for Cache {
+    async fn get_block(&self, position: &BlockPos) -> &'static Block {
+        GenerationCache::get_block_state(self, &position.0).to_block()
+    }
+
+    async fn get_block_state(&self, position: &BlockPos) -> &'static BlockState {
+        GenerationCache::get_block_state(self, &position.0).to_state()
+    }
+
+    async fn get_block_and_state(
+        &self,
+        position: &BlockPos,
+    ) -> (&'static Block, &'static BlockState) {
+        let id = GenerationCache::get_block_state(self, &position.0);
+        (id.to_block(), id.to_state())
+    }
+}
+
+impl GenerationCache for Cache {
+    fn get_center_chunk_mut(&mut self) -> &mut ProtoChunk {
+        let mid = ((self.size * self.size) >> 1) as usize;
+        match &mut self.chunks[mid] {
+            Chunk::Proto(chunk) => chunk.proto_chunk_mut(),
+            _ => panic!(),
+        }
+    }
+
+    fn get_block_state(&self, pos: &Vector3<i32>) -> RawBlockState {
+        let dx = (pos.x >> 4) - self.x;
+        let dz = (pos.z >> 4) - self.y;
+        debug_assert!(dx < self.size && dz < self.size);
+        match &self.chunks[(dx * self.size + dz) as usize] {
+            Chunk::Level(data) => {
+                let chunk = self.rt.block_on(data.read());
+                RawBlockState(
+                    chunk
+                        .section
+                        .get_block_absolute_y(
+                            (pos.x & 15) as usize,
+                            pos.y - chunk.section.min_y,
+                            (pos.z & 15) as usize,
+                        )
+                        .unwrap_or(0),
+                )
+            }
+            Chunk::Proto(data) => data.proto_chunk().get_block_state(pos),
+        }
+    }
+    fn set_block_state(&mut self, pos: &Vector3<i32>, block_state: &BlockState) {
+        let dx = (pos.x >> 4) - self.x;
+        let dz = (pos.z >> 4) - self.y;
+        debug_assert!(dx < self.size && dz < self.size);
+        match &mut self.chunks[(dx * self.size + dz) as usize] {
+            Chunk::Level(data) => {
+                let mut chunk = self.rt.block_on(data.write());
+                let min_y = chunk.section.min_y;
+                chunk.section.set_block_absolute_y(
+                    (pos.x & 15) as usize,
+                    pos.y - min_y,
+                    (pos.z & 15) as usize,
+                    block_state.id,
+                );
+            }
+            Chunk::Proto(data) => {
+                data.proto_chunk_mut().set_block_state(pos, block_state);
+            }
+        }
+    }
+
+    fn get_top_y(&self, heightmap: &HeightMap, pos: &Vector2<i32>) -> i32 {
+        match heightmap {
+            HeightMap::WorldSurfaceWg => self.top_block_height_exclusive(pos),
+            HeightMap::WorldSurface => self.top_block_height_exclusive(pos),
+            HeightMap::OceanFloorWg => self.ocean_floor_height_exclusive(pos),
+            HeightMap::OceanFloor => self.ocean_floor_height_exclusive(pos),
+            HeightMap::MotionBlocking => self.top_motion_blocking_block_height_exclusive(pos),
+            HeightMap::MotionBlockingNoLeaves => {
+                self.top_motion_blocking_block_no_leaves_height_exclusive(pos)
+            }
+        }
+    }
+
+    fn top_motion_blocking_block_height_exclusive(&self, pos: &Vector2<i32>) -> i32 {
+        let dx = (pos.x >> 4) - self.x;
+        let dy = (pos.y >> 4) - self.y;
+        debug_assert!(dx < self.size && dy < self.size);
+        match &self.chunks[(dx * self.size + dy) as usize] {
+            Chunk::Level(data) => {
+                let chunk = self.rt.block_on(data.read());
+                chunk.heightmap.get_height(
+                    ChunkHeightmapType::MotionBlocking,
+                    pos.x,
+                    pos.y,
+                    chunk.section.min_y,
+                )
+            }
+            Chunk::Proto(data) => data
+                .proto_chunk()
+                .top_motion_blocking_block_height_exclusive(pos),
+        }
+    }
+
+    fn top_motion_blocking_block_no_leaves_height_exclusive(&self, pos: &Vector2<i32>) -> i32 {
+        let dx = (pos.x >> 4) - self.x;
+        let dy = (pos.y >> 4) - self.y;
+        debug_assert!(dx < self.size && dy < self.size);
+        match &self.chunks[(dx * self.size + dy) as usize] {
+            Chunk::Level(data) => {
+                let chunk = self.rt.block_on(data.read());
+                chunk.heightmap.get_height(
+                    ChunkHeightmapType::MotionBlockingNoLeaves,
+                    pos.x,
+                    pos.y,
+                    chunk.section.min_y,
+                )
+            }
+            Chunk::Proto(data) => data
+                .proto_chunk()
+                .top_motion_blocking_block_no_leaves_height_exclusive(pos),
+        }
+    }
+
+    fn top_block_height_exclusive(&self, pos: &Vector2<i32>) -> i32 {
+        let dx = (pos.x >> 4) - self.x;
+        let dy = (pos.y >> 4) - self.y;
+        debug_assert!(dx < self.size && dy < self.size);
+        match &self.chunks[(dx * self.size + dy) as usize] {
+            Chunk::Level(data) => {
+                let chunk = self.rt.block_on(data.read());
+                chunk.heightmap.get_height(
+                    ChunkHeightmapType::WorldSurface,
+                    pos.x,
+                    pos.y,
+                    chunk.section.min_y,
+                ) // can we return this?
+            }
+            Chunk::Proto(data) => data.proto_chunk().top_block_height_exclusive(pos),
+        }
+    }
+
+    fn ocean_floor_height_exclusive(&self, pos: &Vector2<i32>) -> i32 {
+        let dx = (pos.x >> 4) - self.x;
+        let dy = (pos.y >> 4) - self.y;
+        debug_assert!(dx < self.size && dy < self.size);
+        match &self.chunks[(dx * self.size + dy) as usize] {
+            Chunk::Level(data) => {
+                0 // todo missing
+            }
+            Chunk::Proto(data) => data.proto_chunk().ocean_floor_height_exclusive(pos),
+        }
+    }
+
+    fn get_biome_for_terrain_gen(&self, global_block_pos: &Vector3<i32>) -> &'static Biome {
+        let dx = (global_block_pos.x >> 4) - self.x;
+        let dy = (global_block_pos.z >> 4) - self.y;
+        debug_assert!(dx < self.size && dy < self.size);
+        match &self.chunks[(dx * self.size + dy) as usize] {
+            Chunk::Level(data) => {
+                // Could this happen?
+                &pumpkin_data::biome::Biome::PLAINS
+            }
+            Chunk::Proto(data) => data
+                .proto_chunk()
+                .get_biome_for_terrain_gen(global_block_pos),
+        }
+    }
+
+    fn is_air(&self, local_pos: &Vector3<i32>) -> bool {
+        GenerationCache::get_block_state(self, local_pos)
+            .to_state()
+            .is_air()
+    }
+}
+
+impl Cache {
+    fn new(x: i32, y: i32, size: i32) -> Cache {
+        let rt = Builder::new_current_thread().enable_all().build().unwrap();
+        Cache {
+            x,
+            y,
+            size,
+            chunks: Vec::with_capacity((size * size) as usize),
+            rt,
+        }
+    }
+    pub fn advance(
+        &mut self,
+        stage: StagedChunkEnum,
+        block_registry: &dyn BlockRegistryExt,
+        settings: &GenerationSettings,
+        random_config: &GlobalRandomConfig,
+        terrain_cache: &TerrainCache,
+        noise_router: &ProtoNoiseRouters,
+        dimension: Dimension,
+    ) {
+        let mid = ((self.size * self.size) >> 1) as usize;
+        match stage {
+            Empty => panic!(),
+            Biomes => match &mut self.chunks[mid] {
+                Chunk::Proto(chunk) => {
+                    chunk.populate_biomes(noise_router, dimension);
+                }
+                _ => panic!(),
+            },
+            Noise => match &mut self.chunks[mid] {
+                Chunk::Proto(chunk) => {
+                    chunk.populate_noise(settings, random_config, noise_router);
+                }
+                _ => panic!(),
+            },
+            Surface => match &mut self.chunks[mid] {
+                Chunk::Proto(chunk) => {
+                    chunk.build_surface(settings, random_config, terrain_cache, noise_router);
+                }
+                _ => panic!(),
+            },
+            Features => {
+                ProtoChunk::generate_features_and_structure(self, block_registry, random_config);
+                match &mut self.chunks[mid] {
+                    Chunk::Proto(chunk) => unsafe {
+                        let data = match ptr::read(chunk) {
+                            StagedChunk::Surface(s) => s,
+                            _ => panic!(),
+                        };
+                        ptr::write(chunk, StagedChunk::Features(data));
+                    },
+                    _ => panic!(),
+                }
+            }
+            StagedChunkEnum::Full => {
+                let ptr = &mut self.chunks[mid];
+                unsafe {
+                    let data = match ptr::read(ptr) {
+                        Chunk::Proto(chunk) => chunk.finalize(settings, dimension).unwrap(),
+                        _ => panic!(),
+                    };
+
+                    ptr::write(ptr, Chunk::Level(Arc::new(RwLock::new(data))));
+                }
+            }
+            _ => panic!(),
+        }
+    }
+}
+
+enum RecvChunk {
+    IO(Chunk),
+    Generation(Cache),
+}
+
+pub struct ChunkListener {
+    single: Mutex<Vec<(ChunkPos, oneshot::Sender<SyncChunk>)>>,
+    global: Mutex<Vec<Sender<(ChunkPos, SyncChunk)>>>,
+}
+impl Default for ChunkListener {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+impl ChunkListener {
+    pub fn new() -> Self {
+        Self {
+            single: Mutex::new(Vec::new()),
+            global: Mutex::new(Vec::new()),
+        }
+    }
+    pub fn add_single_chunk_listener(&self, pos: ChunkPos) -> oneshot::Receiver<SyncChunk> {
+        let (tx, rx) = oneshot::channel();
+        self.single.lock().unwrap().push((pos, tx));
+        rx
+    }
+    pub fn add_global_chunk_listener(&self) -> Receiver<(ChunkPos, SyncChunk)> {
+        let (tx, rx) = crossbeam::channel::unbounded();
+        self.global.lock().unwrap().push(tx);
+        rx
+    }
+    fn process_new_chunk(&self, pos: ChunkPos, chunk: &SyncChunk) {
+        {
+            let mut single = self.single.lock().unwrap();
+            let mut i = 0;
+            let mut len = single.len();
+            while i < len {
+                if single[i].0 == pos {
+                    let (_, send) = single.remove(i);
+                    let _ = send.send(chunk.clone());
+                    log::debug!("single listener send {pos:?}");
+                    len -= 1;
+                    continue;
+                }
+                if single[i].1.is_closed() {
+                    let listener_pos = single[i].0;
+                    single.remove(i);
+                    log::debug!("single dropped send {listener_pos:?}");
+                    len -= 1;
+                    continue;
+                }
+                i += 1;
+            }
+        }
+        {
+            let mut global = self.global.lock().unwrap();
+            let mut i = 0;
+            let mut len = global.len();
+            while i < len {
+                match global[i].send((pos, chunk.clone())) {
+                    Ok(_) => {}
+                    Err(_) => {
+                        log::debug!("one global listener dropped");
+                        global.remove(i);
+                        len -= 1;
+                        continue;
+                    }
+                }
+                i += 1;
+            }
+        }
+    }
+}
+
+pub struct GenerationSchedule {
+    pub queue: Vec<(ChunkPos, i8, StagedChunkEnum)>,
+    pub last_level: ChunkLevel,
+    pub send_level: Arc<LevelChannel>,
+    pub loaded_chunks: Arc<DashMap<Vector2<i32>, SyncChunk>>,
+    pub proto_chunks: HashMapType<ChunkPos, StagedChunk>,
+    pub unload_chunks: HashMapType<ChunkPos, Chunk>,
+    pub occupied: HashSet<ChunkPos>,
+    pub task_mark: HashMapType<ChunkPos, u8>,
+    pub running_task_count: u16,
+    pub recv_chunk: Receiver<(ChunkPos, RecvChunk)>,
+    pub io: Sender<ChunkPos>,
+    pub generate: Sender<(ChunkPos, Cache, StagedChunkEnum)>,
+    pub listener: Arc<ChunkListener>,
+}
+
+impl GenerationSchedule {
+    pub fn create(
+        tracker: &TaskTracker,
+        oi_read_thread_count: usize,
+        gen_thread_count: usize,
+        level: Arc<Level>,
+        level_channel: Arc<LevelChannel>,
+        listener: Arc<ChunkListener>,
+        thread_tracker: &mut Vec<JoinHandle<()>>,
+    ) {
+        let (send_chunk, recv_chunk) = crossbeam::channel::unbounded();
+        let (send_io, recv_io) = crossbeam::channel::bounded(4);
+        let (send_gen, recv_gen) = crossbeam::channel::bounded(20);
+        for _ in 0..oi_read_thread_count {
+            tracker.spawn(Self::io_work(
+                recv_io.clone(),
+                send_chunk.clone(),
+                level.clone(),
+            ));
+        }
+        for i in 0..gen_thread_count {
+            let recv_gen = recv_gen.clone();
+            let send_chunk = send_chunk.clone();
+            let level = level.clone();
+            let builder = thread::Builder::new().name(format!("Generation Thread {i}"));
+            thread_tracker.push(
+                builder
+                    .spawn(move || {
+                        Self::generation_work(recv_gen, send_chunk, level);
+                    })
+                    .unwrap(),
+            );
+        }
+
+        let builder = thread::Builder::new().name("Schedule Thread".to_string());
+        thread_tracker.push(
+            builder
+                .spawn(move || {
+                    Self {
+                        queue: Vec::new(),
+                        last_level: ChunkLevel::default(),
+                        send_level: level_channel,
+                        loaded_chunks: level.loaded_chunks.clone(),
+                        proto_chunks: HashMapType::default(),
+                        unload_chunks: HashMapType::default(),
+                        occupied: HashSet::new(),
+                        task_mark: HashMapType::default(),
+                        running_task_count: 0,
+                        recv_chunk,
+                        io: send_io,
+                        generate: send_gen,
+                        listener,
+                    }
+                    .work(level);
+                })
+                .unwrap(),
+        )
+    }
+
+    fn get_chunk(
+        loaded_chunks: &Arc<DashMap<Vector2<i32>, SyncChunk>>,
+        proto_chunks: &mut HashMapType<ChunkPos, StagedChunk>,
+        pos: ChunkPos,
+    ) -> Option<Chunk> {
+        if let Some(data) = loaded_chunks.get(&pos) {
+            Some(Chunk::Level(data.clone()))
+        } else {
+            proto_chunks.remove(&pos).map(Chunk::Proto)
+        }
+    }
+
+    fn get_chunk_stage_id(
+        loaded_chunks: &Arc<DashMap<Vector2<i32>, SyncChunk>>,
+        proto_chunks: &HashMapType<ChunkPos, StagedChunk>,
+        pos: ChunkPos,
+    ) -> u8 {
+        if loaded_chunks.contains_key(&pos) {
+            6
+        } else if let Some(data) = proto_chunks.get(&pos) {
+            data.stage_id()
+        } else {
+            0
+        }
+    }
+    fn add_chunk(
+        loaded_chunks: &Arc<DashMap<Vector2<i32>, SyncChunk>>,
+        proto_chunks: &mut HashMapType<ChunkPos, StagedChunk>,
+        pos: ChunkPos,
+        chunk: Chunk,
+    ) {
+        match chunk {
+            Chunk::Level(data) => {
+                loaded_chunks.insert(pos, data);
+            }
+            Chunk::Proto(data) => {
+                proto_chunks.insert(pos, data);
+            }
+        }
+    }
+
+    fn resort_work(&mut self, new_level: ChunkLevel) {
+        for pos in self.last_level.keys() {
+            if !new_level.contains_key(pos)
+                && let Some(chunk) =
+                    Self::get_chunk(&self.loaded_chunks, &mut self.proto_chunks, *pos)
+            {
+                self.unload_chunks.insert(*pos, chunk);
+            }
+        }
+        for (pos, level) in &new_level {
+            let old_level = *self.last_level.get(pos).unwrap_or(&ChunkLoading::MAX_LEVEL);
+            if old_level == ChunkLoading::MAX_LEVEL
+                && let Some(chunk) = self.unload_chunks.remove(pos)
+            {
+                Self::add_chunk(&self.loaded_chunks, &mut self.proto_chunks, *pos, chunk);
+            }
+            if old_level != *level && !self.loaded_chunks.contains_key(pos) {
+                let next_stage = StagedChunkEnum::level_to_stage(old_level);
+                let new_highest_stage = StagedChunkEnum::level_to_stage(*level);
+                if next_stage == new_highest_stage {
+                    continue;
+                }
+                if let Some(chunk) = self.proto_chunks.get(pos)
+                    && chunk.stage_id() >= (new_highest_stage as u8)
+                {
+                    continue;
+                }
+                match self.task_mark.entry(*pos) {
+                    Entry::Occupied(mut entry) => {
+                        let mark = entry.get_mut();
+                        for i in (next_stage as u8 + 1)..=(new_highest_stage as u8) {
+                            if (*mark >> i & 1) == 0 {
+                                // no task before
+                                self.queue.push((*pos, i8::MAX, i.into()));
+                                *mark |= 1 << i;
+                            }
+                        }
+                    }
+                    Entry::Vacant(entry) => {
+                        let mut mark = 0;
+                        for i in (next_stage as u8 + 1)..=(new_highest_stage as u8) {
+                            self.queue.push((*pos, i8::MAX, i.into()));
+                            mark |= 1 << i;
+                        }
+                        entry.insert(mark);
+                    }
+                };
+            }
+        }
+        for (pos, level, _) in self.queue.iter_mut() {
+            *level = *new_level.get(pos).unwrap();
+        }
+        self.queue
+            .sort_unstable_by(|(l_pos, l_level, l_stage), (r_pos, r_level, r_stage)| {
+                if l_level != r_level {
+                    l_level.cmp(r_level)
+                } else {
+                    l_stage.cmp(r_stage)
+                }
+            });
+        self.last_level = new_level;
+    }
+    fn try_unload() {
+        // TODO
+    }
+
+    async fn io_work(
+        recv: Receiver<ChunkPos>,
+        send: Sender<(ChunkPos, RecvChunk)>,
+        level: Arc<Level>,
+    ) {
+        log::info!("io thread start");
+        use crate::biome::hash_seed;
+        let biome_mixer_seed = hash_seed(level.world_gen.random_config.seed);
+        let (t_send, mut t_recv) = tokio::sync::mpsc::channel(2);
+        while let Ok(pos) = task::block_in_place(|| recv.recv()) {
+            debug!("io thread receive chunk pos {pos:?}");
+            level
+                .chunk_saver
+                .fetch_chunks(&level.level_folder, &[pos], t_send.clone())
+                .await;
+            if let Some(Loaded(chunk)) = t_recv.recv().await {
+                if send
+                    .send((pos, RecvChunk::IO(Chunk::Level(chunk))))
+                    .is_err()
+                {
+                    log::info!("io thread stop");
+                    break;
+                }
+            } else if send
+                .send((
+                    pos,
+                    RecvChunk::IO(Chunk::Proto(StagedChunk::new(
+                        pos,
+                        gen_settings_from_dimension(&level.world_gen.dimension),
+                        level.world_gen.default_block,
+                        biome_mixer_seed,
+                    ))),
+                ))
+                .is_err()
+            {
+                log::info!("io thread stop");
+                break;
+            }
+        }
+    }
+
+    fn generation_work(
+        recv: Receiver<(ChunkPos, Cache, StagedChunkEnum)>,
+        send: Sender<(ChunkPos, RecvChunk)>,
+        level: Arc<Level>,
+    ) {
+        log::info!(
+            "generation thread start id: {:?} name: {}",
+            thread::current().id(),
+            thread::current().name().unwrap_or("unknown")
+        );
+
+        let settings = gen_settings_from_dimension(&level.world_gen.dimension);
+        while let Ok((pos, mut cache, stage)) = recv.recv() {
+            debug!("generation thread receive chunk pos {pos:?} to stage {stage:?}");
+            cache.advance(
+                stage,
+                level.block_registry.as_ref(),
+                settings,
+                &level.world_gen.random_config,
+                &level.world_gen.terrain_cache,
+                &level.world_gen.base_router,
+                level.world_gen.dimension,
+            );
+            if send.send((pos, RecvChunk::Generation(cache))).is_err() {
+                log::info!(
+                    "generation thread stop id: {:?} name: {}",
+                    thread::current().id(),
+                    thread::current().name().unwrap_or("unknown")
+                );
+                break;
+            }
+        }
+    }
+
+    fn drop_mark(&mut self, stage: StagedChunkEnum, pos: ChunkPos) {
+        let mark = self.task_mark.get_mut(&pos).unwrap();
+        debug_assert!((*mark >> (stage as u8) & 1) == 1);
+        *mark -= 1 << (stage as u8);
+        if *mark == 0 {
+            self.task_mark.remove(&pos);
+        }
+    }
+
+    fn receive_chunk(&mut self, pos: ChunkPos, data: RecvChunk) {
+        debug!("receive chunk pos {pos:?}");
+        match data {
+            RecvChunk::IO(chunk) => match chunk {
+                Chunk::Level(data) => {
+                    self.loaded_chunks.insert(pos, data.clone());
+                    self.task_mark.remove(&pos);
+                    self.listener.process_new_chunk(pos, &data);
+                }
+                Chunk::Proto(data) => {
+                    self.proto_chunks.insert(pos, data);
+                    // i think we don't need to remove mark here
+                }
+            },
+            RecvChunk::Generation(data) => {
+                let mut dx = 0;
+                let mut dy = 0;
+                for chunk in data.chunks {
+                    let new_pos = ChunkPos::new(data.x + dx, data.y + dy);
+                    match chunk {
+                        Chunk::Level(chunk) => {
+                            if new_pos == pos {
+                                // other chunk is borrowed by arc. don't need to return
+                                self.loaded_chunks.insert(new_pos, chunk.clone());
+                                self.task_mark.remove(&new_pos);
+                                self.listener.process_new_chunk(new_pos, &chunk);
+                            }
+                        }
+                        Chunk::Proto(chunk) => {
+                            self.proto_chunks.insert(new_pos, chunk);
+                        }
+                    }
+                    self.occupied.remove(&new_pos);
+                    dy += 1;
+                    if dy == data.size {
+                        dy = 0;
+                        dx += 1;
+                    }
+                }
+            }
+        }
+        self.running_task_count -= 1;
+    }
+
+    fn dump_debug_info(&self, sx: i32, tx: i32, sy: i32, ty: i32) {
+        debug!("queue len {}", self.queue.len());
+        // debug!("queue {:?}", self.queue);
+        debug!("running tasks {}", self.running_task_count);
+        debug!(
+            "global listener count {}",
+            self.listener.global.lock().unwrap().len()
+        );
+        debug!(
+            "single listener count {}",
+            self.listener.single.lock().unwrap().len()
+        );
+        let mut s = String::new();
+        for x in sx..=tx {
+            for y in sy..=ty {
+                s += Self::get_chunk_stage_id(
+                    &self.loaded_chunks,
+                    &self.proto_chunks,
+                    ChunkPos::new(x, y),
+                )
+                .to_string()
+                .as_str();
+                s += " ";
+            }
+            s += "\n";
+        }
+        debug!("chunk stage:\n{s}\n");
+
+        ChunkLoading::dump_level_debug(&self.last_level, sx, tx, sy, ty);
+    }
+
+    fn work(mut self, level: Arc<Level>) {
+        log::info!(
+            "schedule thread start id: {:?} name: {}",
+            thread::current().id(),
+            thread::current().name().unwrap_or("unknown")
+        );
+        if let Some(new_level) = self.send_level.wait_and_get(&level) {
+            debug!("receive new level");
+            self.resort_work(new_level);
+        }
+        let mut clock = Instant::now();
+        loop {
+            if level.is_shutting_down.load(Relaxed) {
+                break;
+            }
+            let mut nothing = true;
+            let mut len = self.queue.len();
+            let mut i = 0;
+
+            let now = Instant::now();
+            if now - clock > Duration::from_secs(5) {
+                self.dump_debug_info(-20, 20, -20, 20);
+                clock = now;
+            }
+            'outer: while i < len {
+                while let Ok((pos, data)) = self.recv_chunk.try_recv() {
+                    self.receive_chunk(pos, data);
+                }
+
+                let (pos, _, stage) = self.queue[i];
+
+                let level = *self
+                    .last_level
+                    .get(&pos)
+                    .unwrap_or(&ChunkLoading::MAX_LEVEL);
+                if level == ChunkLoading::MAX_LEVEL {
+                    self.drop_mark(stage, pos);
+                    self.queue.remove(i);
+                    len -= 1;
+                    continue;
+                }
+
+                let highest_stage = StagedChunkEnum::level_to_stage(level);
+                if (highest_stage as u8) < (stage as u8) {
+                    self.drop_mark(stage, pos);
+                    self.queue.remove(i);
+                    len -= 1;
+                    continue;
+                }
+
+                let current_stage =
+                    Self::get_chunk_stage_id(&self.loaded_chunks, &mut self.proto_chunks, pos);
+                if current_stage >= (stage as u8) {
+                    self.drop_mark(stage, pos);
+                    self.queue.remove(i);
+                    len -= 1;
+                    continue;
+                }
+
+                if stage == Empty {
+                    nothing = false;
+                    self.running_task_count += 1;
+                    self.io.send(pos).expect("oi thread close unexpectedly");
+                    self.queue.remove(i);
+                    len -= 1;
+                    continue;
+                }
+
+                let radius = stage.get_radius();
+                let depend = stage.get_dependencies();
+                for dx in -radius..=radius {
+                    for dy in -radius..=radius {
+                        let new_pos = pos.add_raw(dx, dy);
+                        let dst = max(abs(dx), abs(dy)) as usize;
+                        if self.occupied.contains(&new_pos)
+                            || Self::get_chunk_stage_id(
+                                &self.loaded_chunks,
+                                &mut self.proto_chunks,
+                                new_pos,
+                            ) < (depend[dst] as u8)
+                        {
+                            i += 1;
+                            continue 'outer;
+                        }
+                    }
+                }
+                let mut cache = Cache::new(pos.x - radius, pos.y - radius, (radius << 1) + 1);
+                for dx in -radius..=radius {
+                    for dy in -radius..=radius {
+                        let new_pos = pos.add_raw(dx, dy);
+                        cache.chunks.push(
+                            Self::get_chunk(&self.loaded_chunks, &mut self.proto_chunks, new_pos)
+                                .expect("Chunk missing"),
+                        );
+                        self.occupied.insert(new_pos);
+                    }
+                }
+                self.running_task_count += 1;
+                self.generate
+                    .send((pos, cache, stage))
+                    .expect("oi thread close unexpectedly");
+                self.queue.remove(i);
+                len -= 1;
+            }
+            if len == 0 {
+                debug!("the queue is empty. thread sleep");
+                'out: while self.running_task_count > 0 {
+                    let (pos, data) = self.recv_chunk.recv().expect("recv_chunk stop");
+                    self.receive_chunk(pos, data);
+                    if let Some(new_level) = self.send_level.get() {
+                        debug!("receive new level");
+                        self.resort_work(new_level);
+                        break 'out;
+                    }
+                }
+                if let Some(new_level) = self.send_level.wait_and_get(&level) {
+                    debug!("receive new level");
+                    self.resort_work(new_level);
+                }
+            } else if let Some(new_level) = self.send_level.get() {
+                debug!("receive new level");
+                self.resort_work(new_level);
+            } else if nothing && self.running_task_count > 0 {
+                debug!("nothing to do. thread sleep.");
+                if let Ok((pos, data)) = self.recv_chunk.recv() {
+                    self.receive_chunk(pos, data);
+                }
+            }
+        }
+    }
+}
