+++ pumpkin-world/src/level.rs
@@ -1,5 +1,5 @@
+use crate::chunk_system::{ChunkListener, ChunkLoading, GenerationSchedule, LevelChannel};
 use crate::generation::generator::VanillaGenerator;
-use crate::level::new_chunk_system::{ChunkLoading, GenerationSchedule, LevelChannel};
 use crate::{
     BlockStateId,
     block::{RawBlockState, entities::BlockEntity},
@@ -44,1112 +44,6 @@ use tokio_util::task::TaskTracker;
 
 pub type SyncChunk = Arc<RwLock<ChunkData>>;
 pub type SyncEntityChunk = Arc<RwLock<ChunkEntityData>>;
-
-pub mod new_chunk_system {
-    use crate::block::RawBlockState;
-    use crate::chunk::ChunkHeightmapType;
-    use crate::chunk::io::FileIO;
-    use crate::chunk::io::LoadedData::Loaded;
-    use crate::dimension::Dimension;
-    
-    use crate::generation::height_limit::HeightLimitView;
-    
-    use crate::generation::proto_chunk::{
-        GenerationCache, StagedChunk, TerrainCache,
-    };
-    use crate::generation::settings::{GenerationSettings, gen_settings_from_dimension};
-    use crate::level::new_chunk_system::StagedChunkEnum::{
-        Biomes, Empty, Features, Noise, Surface,
-    };
-    use crate::level::{Level, SyncChunk};
-    use crate::world::{BlockAccessor, BlockRegistryExt};
-    use crate::{GlobalRandomConfig, ProtoChunk, ProtoNoiseRouters};
-    use async_trait::async_trait;
-    use crossbeam::channel::{Receiver, Sender};
-    use dashmap::DashMap;
-    use itertools::Itertools;
-    use log::debug;
-    use num_traits::abs;
-    use pumpkin_data::biome::Biome;
-    
-    
-    use pumpkin_data::{Block, BlockState};
-    use pumpkin_util::HeightMap;
-    use pumpkin_util::math::position::BlockPos;
-    use pumpkin_util::math::vector2::Vector2;
-    use pumpkin_util::math::vector3::Vector3;
-    
-    use std::cmp::{PartialEq, max, min};
-    use std::collections::hash_map::Entry;
-    use std::collections::{HashMap, HashSet, VecDeque};
-    use std::mem::swap;
-    use std::sync::{Arc, Condvar, Mutex};
-    
-    use std::time::{Duration, Instant};
-    use std::{ptr, thread};
-    use tokio::runtime::{Builder, Runtime};
-    use tokio::sync::RwLock;
-    use tokio::task;
-    
-    use tokio_util::task::TaskTracker;
-
-    type ChunkPos = Vector2<i32>;
-    type ChunkLevel = HashMap<ChunkPos, i8>;
-
-    pub struct ChunkLoading {
-        pub pos_level: ChunkLevel,
-        pub ticket: HashMap<ChunkPos, Vec<i8>>, // TODO lifetime & id
-    }
-
-    impl Default for ChunkLoading {
-        fn default() -> Self {
-            Self::new()
-        }
-    }
-
-    impl ChunkLoading {
-        pub fn new() -> Self {
-            Self {
-                pos_level: ChunkLevel::new(),
-                ticket: HashMap::new(),
-            }
-        }
-        pub const MAX_LEVEL: i8 = 46; // 46 就已经卸载了，不能是46
-        fn run_decrease_update(pos_level: &mut ChunkLevel, pos: ChunkPos, level: i8) {
-            debug_assert!(level < Self::MAX_LEVEL);
-            debug_assert!(level < *pos_level.get(&pos).unwrap_or(&Self::MAX_LEVEL));
-            let mut queue: VecDeque<(ChunkPos, i8)> = Default::default();
-            queue.push_back((pos, level));
-            while !queue.is_empty() {
-                let (pos, level) = queue.pop_front().unwrap(); // can use unsafe
-                debug_assert!(level < Self::MAX_LEVEL);
-                match pos_level.entry(pos) {
-                    Entry::Occupied(mut entry) => {
-                        let old = entry.get_mut();
-                        if *old <= level {
-                            continue;
-                        }
-                        *old = level;
-                    }
-                    Entry::Vacant(empty) => {
-                        empty.insert(level);
-                    }
-                }
-                let spread_level = level + 1;
-                if spread_level >= Self::MAX_LEVEL {
-                    continue;
-                }
-                for dx in -1..2 {
-                    for dy in -1..2 {
-                        let new_pos = pos.add_raw(dx, dy);
-                        if new_pos != pos
-                            && spread_level < *pos_level.get(&new_pos).unwrap_or(&Self::MAX_LEVEL)
-                        {
-                            queue.push_back((new_pos, spread_level));
-                        }
-                    }
-                }
-            }
-        }
-
-        fn get_level_from_neighbour(&self, pos: ChunkPos) -> i8 {
-            let mut ret = Self::MAX_LEVEL;
-            for dx in -1..2 {
-                for dy in -1..2 {
-                    let new_pos = pos.add_raw(dx, dy);
-                    ret = min(
-                        *self.pos_level.get(&new_pos).unwrap_or(&Self::MAX_LEVEL) + 1,
-                        ret,
-                    );
-                }
-            }
-            ret
-        }
-
-        fn run_increase_update(&mut self, pos: ChunkPos, level: i8) {
-            // TODO super slow
-            debug_assert!(level < Self::MAX_LEVEL);
-            let range = Self::MAX_LEVEL - level - 1;
-            for dx in -range..range {
-                for dy in -range..range {
-                    let new_pos = pos.add_raw(dx as i32, dy as i32);
-                    self.pos_level.remove(&new_pos);
-                }
-            }
-            for dst in (1..range).rev() {
-                let min_x = pos.x - dst as i32;
-                let max_x = pos.x - dst as i32;
-                let min_y = pos.y - dst as i32;
-                let max_y = pos.y - dst as i32;
-                for y in min_y..max_y {
-                    let new_pos = ChunkPos::new(max_x, y);
-                    let new_level = self.get_level_from_neighbour(new_pos);
-                    if new_level < Self::MAX_LEVEL {
-                        self.pos_level.insert(new_pos, new_level);
-                    }
-                }
-                for x in min_x..max_x {
-                    let new_pos = ChunkPos::new(x, max_y);
-                    let new_level = self.get_level_from_neighbour(new_pos);
-                    if new_level < Self::MAX_LEVEL {
-                        self.pos_level.insert(new_pos, new_level);
-                    }
-                }
-                for y in (min_y + 1)..=max_y {
-                    let new_pos = ChunkPos::new(min_x, y);
-                    let new_level = self.get_level_from_neighbour(new_pos);
-                    if new_level < Self::MAX_LEVEL {
-                        self.pos_level.insert(new_pos, new_level);
-                    }
-                }
-                for x in (min_x + 1)..=max_x {
-                    let new_pos = ChunkPos::new(x, min_y);
-                    let new_level = self.get_level_from_neighbour(new_pos);
-                    if new_level < Self::MAX_LEVEL {
-                        self.pos_level.insert(new_pos, new_level);
-                    }
-                }
-            }
-            let new_level = self.get_level_from_neighbour(pos);
-            if new_level < Self::MAX_LEVEL {
-                self.pos_level.insert(pos, new_level);
-            }
-            for (ticket_pos, levels) in &self.ticket {
-                if abs(ticket_pos.x - pos.x) <= range as i32
-                    && abs(ticket_pos.y - pos.y) <= range as i32
-                {
-                    Self::run_decrease_update(
-                        &mut self.pos_level,
-                        pos,
-                        *levels.iter().max().unwrap(),
-                    );
-                }
-            }
-        }
-        pub fn add_ticket(&mut self, pos: ChunkPos, level: i8) {
-            debug_assert!(level < Self::MAX_LEVEL);
-            let old_level = *self.pos_level.get(&pos).unwrap_or(&Self::MAX_LEVEL);
-            use std::collections::hash_map::Entry;
-            match self.ticket.entry(pos) {
-                Entry::Occupied(mut vec) => {
-                    vec.get_mut().push(level);
-                }
-                Entry::Vacant(empty) => {
-                    empty.insert(vec![level]);
-                }
-            }
-            if level < old_level {
-                ChunkLoading::run_decrease_update(&mut self.pos_level, pos, level);
-            }
-        }
-        pub fn remove_ticket(&mut self, pos: ChunkPos, level: i8) {
-            debug_assert!(level < Self::MAX_LEVEL);
-            let old_level = *self.pos_level.get(&pos).unwrap_or(&Self::MAX_LEVEL);
-            let Some(vec) = self.ticket.get_mut(&pos) else {
-                log::warn!("No ticket found at {pos:?}");
-                return;
-            };
-            vec.remove(vec.iter().find_position(|x| **x == level).unwrap().0);
-            let source = *vec.iter().max().unwrap_or(&Self::MAX_LEVEL);
-            if vec.is_empty() {
-                self.ticket.remove(&pos);
-            }
-            if level == old_level && source != level {
-                self.run_increase_update(pos, old_level);
-            }
-        }
-        pub fn get_cloned_level(&self) -> ChunkLevel {
-            self.pos_level.clone()
-        }
-    }
-
-    #[repr(u8)]
-    #[derive(Debug, Copy, Clone, Eq, Ord, PartialEq, PartialOrd)]
-    pub enum StagedChunkEnum {
-        None,
-        /// Initial empty chunk, ready for biome population
-        Empty = 1, // EMPTY STRUCTURE_STARTS STRUCTURE_REFERENCES
-        /// Chunk with biomes populated, ready for noise generation
-        Biomes,
-        /// Chunk with terrain noise generated, ready for surface building
-        Noise,
-        /// Chunk with surface built, ready for features and structures
-        Surface, // SURFACE CARVERS
-        /// Chunk with features and structures, ready for finalization
-        Features, // FEATURES INITIALIZE_LIGHT LIGHT SPAWN
-        /// Fully generated chunk
-        Full,
-    }
-
-    impl From<u8> for StagedChunkEnum {
-        fn from(v: u8) -> Self {
-            match v {
-                1 => Self::Empty,
-                2 => Self::Biomes,
-                3 => Self::Noise,
-                4 => Self::Surface,
-                5 => Self::Features,
-                6 => Self::Full,
-                _ => panic!(),
-            }
-        }
-    }
-    impl StagedChunkEnum {
-        const MAX_STAGE_ID: u8 = Self::Full as u8;
-        fn level_to_stage(level: i8) -> Self {
-            if level <= 33 {
-                Self::Full
-            } else if level <= 35 {
-                Self::Features
-            } else if level <= 36 {
-                Self::Surface
-            } else if level <= 37 {
-                Self::Biomes
-            } else if level <= 45 {
-                Self::Empty
-            } else {
-                Self::None
-            }
-        }
-        fn get_radius(self) -> i32 {
-            // 这个是不包括自己的范围
-            match self {
-                StagedChunkEnum::Empty => 0,
-                StagedChunkEnum::Biomes => 8,
-                StagedChunkEnum::Noise => 9,
-                StagedChunkEnum::Surface => 9,
-                StagedChunkEnum::Features => 10,
-                StagedChunkEnum::Full => 11,
-                _ => panic!(),
-            }
-        }
-        fn get_dependencies(self) -> &'static [StagedChunkEnum] {
-            match self {
-                StagedChunkEnum::Empty => &[],
-                StagedChunkEnum::Biomes => &[
-                    Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty,
-                ],
-                StagedChunkEnum::Noise => &[
-                    Biomes, Biomes, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty,
-                ],
-                StagedChunkEnum::Surface => &[
-                    Noise, Biomes, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty,
-                ],
-                StagedChunkEnum::Features => &[
-                    Surface, Surface, Biomes, Empty, Empty, Empty, Empty, Empty, Empty, Empty,
-                    Empty,
-                ],
-                StagedChunkEnum::Full => &[
-                    Features, Features, Surface, Biomes, Empty, Empty, Empty, Empty, Empty, Empty,
-                    Empty, Empty,
-                ],
-                _ => panic!(),
-            }
-        }
-    }
-
-    pub struct LevelChannel {
-        pub value: Mutex<Option<ChunkLevel>>,
-        pub notify: Condvar,
-    }
-
-    impl Default for LevelChannel {
-        fn default() -> Self {
-            Self::new()
-        }
-    }
-
-    impl LevelChannel {
-        pub fn new() -> Self {
-            Self {
-                value: Mutex::new(None),
-                notify: Condvar::new(),
-            }
-        }
-
-        pub fn set(&self, new_value: ChunkLevel) {
-            *self.value.lock().unwrap() = Some(new_value);
-            self.notify.notify_one();
-        }
-        pub fn get(&self) -> Option<ChunkLevel> {
-            let mut lock = self.value.lock().unwrap();
-            let mut ret = None;
-            swap(&mut ret, &mut *lock);
-            ret
-        }
-        pub fn wait_and_get(&self) -> Option<ChunkLevel> {
-            // TODO if this return None, stop thread
-            let mut lock = self.value.lock().unwrap();
-            while lock.is_none() {
-                lock = self.notify.wait(lock).unwrap();
-            }
-            let mut ret = None;
-            swap(&mut ret, &mut *lock);
-            ret
-        }
-        pub fn notify(&self) {
-            // TODO ensure thread stop
-            self.notify.notify_one();
-        }
-    }
-
-    pub enum Chunk {
-        Level(SyncChunk),
-        Proto(StagedChunk),
-    }
-
-    impl Chunk {
-        fn to_stage_id(chunk: &Option<Self>) -> u8 {
-            match chunk {
-                None => 0,
-                Some(Chunk::Proto(data)) => data.stage_id(),
-                Some(Chunk::Level(_)) => 6,
-            }
-        }
-    }
-
-    struct Cache {
-        x: i32,
-        y: i32,
-        size: i32,
-        pub chunks: Vec<Chunk>,
-        rt: Runtime,
-    }
-
-    impl HeightLimitView for Cache {
-        fn height(&self) -> u16 {
-            let mid = ((self.size * self.size) >> 1) as usize;
-            match &self.chunks[mid] {
-                Chunk::Proto(chunk) => chunk.proto_chunk().height(),
-                _ => panic!(),
-            }
-        }
-
-        fn bottom_y(&self) -> i8 {
-            let mid = ((self.size * self.size) >> 1) as usize;
-            match &self.chunks[mid] {
-                Chunk::Proto(chunk) => chunk.proto_chunk().bottom_y(),
-                _ => panic!(),
-            }
-        }
-    }
-
-    #[async_trait]
-    impl BlockAccessor for Cache {
-        async fn get_block(&self, position: &BlockPos) -> &'static Block {
-            GenerationCache::get_block_state(self, &position.0).to_block()
-        }
-
-        async fn get_block_state(&self, position: &BlockPos) -> &'static BlockState {
-            GenerationCache::get_block_state(self, &position.0).to_state()
-        }
-
-        async fn get_block_and_state(
-            &self,
-            position: &BlockPos,
-        ) -> (&'static Block, &'static BlockState) {
-            let id = GenerationCache::get_block_state(self, &position.0);
-            (id.to_block(), id.to_state())
-        }
-    }
-
-    impl GenerationCache for Cache {
-        fn get_center_chunk_mut(&mut self) -> &mut ProtoChunk {
-            let mid = ((self.size * self.size) >> 1) as usize;
-            match &mut self.chunks[mid] {
-                Chunk::Proto(chunk) => chunk.proto_chunk_mut(),
-                _ => panic!(),
-            }
-        }
-
-        fn get_block_state(&self, pos: &Vector3<i32>) -> RawBlockState {
-            let dx = (pos.x >> 4) - self.x;
-            let dz = (pos.z >> 4) - self.y;
-            debug_assert!(dx < self.size && dz < self.size);
-            match &self.chunks[(dx * self.size + dz) as usize] {
-                Chunk::Level(data) => {
-                    let chunk = self.rt.block_on(data.read());
-                    RawBlockState(
-                        chunk
-                            .section
-                            .get_block_absolute_y(
-                                (pos.x & 15) as usize,
-                                pos.y - chunk.section.min_y,
-                                (pos.z & 15) as usize,
-                            )
-                            .unwrap_or(0),
-                    )
-                }
-                Chunk::Proto(data) => data.proto_chunk().get_block_state(pos),
-            }
-        }
-        fn set_block_state(&mut self, pos: &Vector3<i32>, block_state: &BlockState) {
-            let dx = (pos.x >> 4) - self.x;
-            let dz = (pos.z >> 4) - self.y;
-            debug_assert!(dx < self.size && dz < self.size);
-            match &mut self.chunks[(dx * self.size + dz) as usize] {
-                Chunk::Level(data) => {
-                    let mut chunk = self.rt.block_on(data.write());
-                    let min_y = chunk.section.min_y;
-                    chunk.section.set_block_absolute_y(
-                        (pos.x & 15) as usize,
-                        pos.y - min_y,
-                        (pos.z & 15) as usize,
-                        block_state.id,
-                    );
-                }
-                Chunk::Proto(data) => {
-                    data.proto_chunk_mut().set_block_state(pos, block_state);
-                }
-            }
-        }
-
-        fn get_top_y(&self, heightmap: &HeightMap, pos: &Vector2<i32>) -> i32 {
-            match heightmap {
-                HeightMap::WorldSurfaceWg => self.top_block_height_exclusive(pos),
-                HeightMap::WorldSurface => self.top_block_height_exclusive(pos),
-                HeightMap::OceanFloorWg => self.ocean_floor_height_exclusive(pos),
-                HeightMap::OceanFloor => self.ocean_floor_height_exclusive(pos),
-                HeightMap::MotionBlocking => self.top_motion_blocking_block_height_exclusive(pos),
-                HeightMap::MotionBlockingNoLeaves => {
-                    self.top_motion_blocking_block_no_leaves_height_exclusive(pos)
-                }
-            }
-        }
-
-        fn top_motion_blocking_block_height_exclusive(&self, pos: &Vector2<i32>) -> i32 {
-            let dx = (pos.x >> 4) - self.x;
-            let dy = (pos.y >> 4) - self.y;
-            debug_assert!(dx < self.size && dy < self.size);
-            match &self.chunks[(dx * self.size + dy) as usize] {
-                Chunk::Level(data) => {
-                    let chunk = self.rt.block_on(data.read());
-                    chunk.heightmap.get_height(
-                        ChunkHeightmapType::MotionBlocking,
-                        pos.x,
-                        pos.y,
-                        chunk.section.min_y,
-                    )
-                }
-                Chunk::Proto(data) => data
-                    .proto_chunk()
-                    .top_motion_blocking_block_height_exclusive(pos),
-            }
-        }
-
-        fn top_motion_blocking_block_no_leaves_height_exclusive(&self, pos: &Vector2<i32>) -> i32 {
-            let dx = (pos.x >> 4) - self.x;
-            let dy = (pos.y >> 4) - self.y;
-            debug_assert!(dx < self.size && dy < self.size);
-            match &self.chunks[(dx * self.size + dy) as usize] {
-                Chunk::Level(data) => {
-                    let chunk = self.rt.block_on(data.read());
-                    chunk.heightmap.get_height(
-                        ChunkHeightmapType::MotionBlockingNoLeaves,
-                        pos.x,
-                        pos.y,
-                        chunk.section.min_y,
-                    )
-                }
-                Chunk::Proto(data) => data
-                    .proto_chunk()
-                    .top_motion_blocking_block_no_leaves_height_exclusive(pos),
-            }
-        }
-
-        fn top_block_height_exclusive(&self, pos: &Vector2<i32>) -> i32 {
-            let dx = (pos.x >> 4) - self.x;
-            let dy = (pos.y >> 4) - self.y;
-            debug_assert!(dx < self.size && dy < self.size);
-            match &self.chunks[(dx * self.size + dy) as usize] {
-                Chunk::Level(data) => {
-                    let chunk = self.rt.block_on(data.read());
-                    chunk.heightmap.get_height(
-                        ChunkHeightmapType::WorldSurface,
-                        pos.x,
-                        pos.y,
-                        chunk.section.min_y,
-                    ) // can we return this?
-                }
-                Chunk::Proto(data) => data.proto_chunk().top_block_height_exclusive(pos),
-            }
-        }
-
-        fn ocean_floor_height_exclusive(&self, pos: &Vector2<i32>) -> i32 {
-            let dx = (pos.x >> 4) - self.x;
-            let dy = (pos.y >> 4) - self.y;
-            debug_assert!(dx < self.size && dy < self.size);
-            match &self.chunks[(dx * self.size + dy) as usize] {
-                Chunk::Level(data) => {
-                    0 // todo missing
-                }
-                Chunk::Proto(data) => data.proto_chunk().ocean_floor_height_exclusive(pos),
-            }
-        }
-
-        fn get_biome_for_terrain_gen(&self, global_block_pos: &Vector3<i32>) -> &'static Biome {
-            let dx = (global_block_pos.x >> 4) - self.x;
-            let dy = (global_block_pos.z >> 4) - self.y;
-            debug_assert!(dx < self.size && dy < self.size);
-            match &self.chunks[(dx * self.size + dy) as usize] {
-                Chunk::Level(data) => {
-                    // Could this happen?
-                    &pumpkin_data::biome::Biome::PLAINS
-                }
-                Chunk::Proto(data) => data
-                    .proto_chunk()
-                    .get_biome_for_terrain_gen(global_block_pos),
-            }
-        }
-
-        fn is_air(&self, local_pos: &Vector3<i32>) -> bool {
-            GenerationCache::get_block_state(self, local_pos)
-                .to_state()
-                .is_air()
-        }
-    }
-
-    impl Cache {
-        fn new(x: i32, y: i32, size: i32) -> Cache {
-            let rt = Builder::new_current_thread().enable_all().build().unwrap();
-            Cache {
-                x,
-                y,
-                size,
-                chunks: Vec::with_capacity((size * size) as usize),
-                rt,
-            }
-        }
-        pub fn advance(
-            &mut self,
-            stage: StagedChunkEnum,
-            block_registry: &dyn BlockRegistryExt,
-            settings: &GenerationSettings,
-            random_config: &GlobalRandomConfig,
-            terrain_cache: &TerrainCache,
-            noise_router: &ProtoNoiseRouters,
-            dimension: Dimension,
-        ) {
-            let mid = ((self.size * self.size) >> 1) as usize;
-            match stage {
-                Empty => panic!(),
-                Biomes => match &mut self.chunks[mid] {
-                    Chunk::Proto(chunk) => {
-                        chunk.populate_biomes(noise_router, dimension);
-                    }
-                    _ => panic!(),
-                },
-                Noise => match &mut self.chunks[mid] {
-                    Chunk::Proto(chunk) => {
-                        chunk.populate_noise(settings, random_config, noise_router);
-                    }
-                    _ => panic!(),
-                },
-                Surface => match &mut self.chunks[mid] {
-                    Chunk::Proto(chunk) => {
-                        chunk.build_surface(settings, random_config, terrain_cache, noise_router);
-                    }
-                    _ => panic!(),
-                },
-                Features => {
-                    ProtoChunk::generate_features_and_structure(
-                        self,
-                        block_registry,
-                        random_config,
-                    );
-                    match &mut self.chunks[mid] {
-                        Chunk::Proto(chunk) => unsafe {
-                            let data = match ptr::read(chunk) {
-                                StagedChunk::Surface(s) => s,
-                                _ => panic!(),
-                            };
-                            ptr::write(chunk, StagedChunk::Features(data));
-                        },
-                        _ => panic!(),
-                    }
-                }
-                StagedChunkEnum::Full => {
-                    let ptr = &mut self.chunks[mid];
-                    unsafe {
-                        let data = match ptr::read(ptr) {
-                            Chunk::Proto(chunk) => chunk.finalize(settings, dimension).unwrap(),
-                            _ => panic!(),
-                        };
-
-                        ptr::write(ptr, Chunk::Level(Arc::new(RwLock::new(data))));
-                    }
-                }
-                _ => panic!(),
-            }
-        }
-    }
-
-    enum RecvChunk {
-        IO(Chunk),
-        Generation(Cache),
-    }
-
-    pub struct GenerationSchedule {
-        pub queue: Vec<(ChunkPos, i8, StagedChunkEnum)>,
-        pub last_level: ChunkLevel,
-        pub send_level: Arc<LevelChannel>,
-        pub loaded_chunks: Arc<DashMap<Vector2<i32>, SyncChunk>>,
-        pub proto_chunks: HashMap<ChunkPos, StagedChunk>,
-        pub unload_chunks: HashMap<ChunkPos, Chunk>,
-        pub occupied: HashSet<ChunkPos>,
-        pub task_mark: HashMap<ChunkPos, u8>,
-        pub running_task_count: u16,
-        pub recv_chunk: Receiver<(ChunkPos, RecvChunk)>,
-        pub io: Sender<ChunkPos>,
-        pub generate: Sender<(ChunkPos, Cache, StagedChunkEnum)>,
-    }
-
-    impl GenerationSchedule {
-        pub(crate) fn new(
-            tracker: &TaskTracker,
-            oi_thread_count: u8,
-            gen_thread_count: u8,
-            level: Arc<Level>,
-        ) -> Arc<LevelChannel> {
-            let (send_chunk, recv_chunk) = crossbeam::channel::unbounded();
-            let (send_io, recv_io) = crossbeam::channel::bounded(4);
-            let (send_gen, recv_gen) = crossbeam::channel::bounded(20);
-            let send_level = Arc::new(LevelChannel::new());
-            for _ in 0..oi_thread_count {
-                tracker.spawn(Self::io_work(
-                    recv_io.clone(),
-                    send_chunk.clone(),
-                    level.clone(),
-                ));
-            }
-            for _ in 0..gen_thread_count {
-                let recv_gen = recv_gen.clone();
-                let send_chunk = send_chunk.clone();
-                let level = level.clone();
-                std::thread::spawn(move || {
-                    Self::generation_work(recv_gen, send_chunk, level);
-                });
-            }
-            let send_level_clone = send_level.clone();
-            std::thread::spawn(move || {
-                Self {
-                    queue: Vec::new(),
-                    last_level: ChunkLevel::new(),
-                    send_level: send_level_clone,
-                    loaded_chunks: level.loaded_chunks.clone(),
-                    proto_chunks: HashMap::new(),
-                    unload_chunks: HashMap::new(),
-                    occupied: HashSet::new(),
-                    task_mark: HashMap::new(),
-                    running_task_count: 0,
-                    recv_chunk,
-                    io: send_io,
-                    generate: send_gen,
-                }
-                .work();
-            });
-            send_level
-        }
-
-        fn get_chunk(
-            loaded_chunks: &Arc<DashMap<Vector2<i32>, SyncChunk>>,
-            proto_chunks: &mut HashMap<ChunkPos, StagedChunk>,
-            pos: ChunkPos,
-        ) -> Option<Chunk> {
-            if let Some(data) = loaded_chunks.get(&pos) {
-                Some(Chunk::Level(data.clone()))
-            } else { proto_chunks.remove(&pos).map(Chunk::Proto) }
-        }
-
-        fn get_chunk_stage_id(
-            loaded_chunks: &Arc<DashMap<Vector2<i32>, SyncChunk>>,
-            proto_chunks: &HashMap<ChunkPos, StagedChunk>,
-            pos: ChunkPos,
-        ) -> u8 {
-            if loaded_chunks.contains_key(&pos) {
-                6
-            } else if let Some(data) = proto_chunks.get(&pos) {
-                data.stage_id()
-            } else {
-                0
-            }
-        }
-        fn add_chunk(
-            loaded_chunks: &Arc<DashMap<Vector2<i32>, SyncChunk>>,
-            proto_chunks: &mut HashMap<ChunkPos, StagedChunk>,
-            pos: ChunkPos,
-            chunk: Chunk,
-        ) {
-            match chunk {
-                Chunk::Level(data) => {
-                    loaded_chunks.insert(pos, data);
-                }
-                Chunk::Proto(data) => {
-                    proto_chunks.insert(pos, data);
-                }
-            }
-        }
-
-        fn resort_work(&mut self, new_level: ChunkLevel) {
-            for pos in self.last_level.keys() {
-                if !new_level.contains_key(pos)
-                    && let Some(chunk) =
-                        Self::get_chunk(&self.loaded_chunks, &mut self.proto_chunks, *pos)
-                    {
-                        self.unload_chunks.insert(*pos, chunk);
-                    }
-            }
-            for (pos, level) in &new_level {
-                let old_level = *self
-                    .last_level
-                    .get(pos)
-                    .unwrap_or(&ChunkLoading::MAX_LEVEL);
-                if old_level == ChunkLoading::MAX_LEVEL
-                    && let Some(chunk) = self.unload_chunks.remove(pos) {
-                        Self::add_chunk(&self.loaded_chunks, &mut self.proto_chunks, *pos, chunk);
-                    }
-                if old_level != *level
-                    && !self.loaded_chunks.contains_key(pos) {
-                        let next_stage = StagedChunkEnum::level_to_stage(old_level);
-                        let new_highest_stage = StagedChunkEnum::level_to_stage(*level);
-                        if next_stage == new_highest_stage {
-                            continue;
-                        }
-                        if let Some(chunk) = self.proto_chunks.get(pos)
-                            && chunk.stage_id() >= (new_highest_stage as u8)
-                        {
-                            continue;
-                        }
-                        match self.task_mark.entry(*pos) {
-                            Entry::Occupied(mut entry) => {
-                                let mark = entry.get_mut();
-                                for i in (next_stage as u8 + 1)..=(new_highest_stage as u8) {
-                                    if (*mark >> i & 1) == 0 {
-                                        // no task before
-                                        self.queue.push((*pos, i8::MAX, i.into()));
-                                        *mark |= 1 << i;
-                                    }
-                                }
-                            }
-                            Entry::Vacant(entry) => {
-                                let mut mark = 0;
-                                for i in (next_stage as u8 + 1)..=(new_highest_stage as u8) {
-                                    self.queue.push((*pos, i8::MAX, i.into()));
-                                    mark |= 1 << i;
-                                }
-                                entry.insert(mark);
-                            }
-                        };
-                    }
-            }
-            for (pos, level, _) in self.queue.iter_mut() {
-                *level = *new_level.get(pos).unwrap();
-            }
-            self.queue
-                .sort_unstable_by(|(l_pos, l_level, l_stage), (r_pos, r_level, r_stage)| {
-                    if l_level != r_level {
-                        l_level.cmp(r_level)
-                    } else {
-                        l_stage.cmp(r_stage)
-                    }
-                });
-            self.last_level = new_level;
-        }
-        fn try_unload() {
-            // TODO
-        }
-
-        async fn io_work(
-            recv: Receiver<ChunkPos>,
-            send: Sender<(ChunkPos, RecvChunk)>,
-            level: Arc<Level>,
-        ) {
-            log::info!("io thread start");
-            use crate::biome::hash_seed;
-            let biome_mixer_seed = hash_seed(level.world_gen.random_config.seed);
-            let (t_send, mut t_recv) = tokio::sync::mpsc::channel(2);
-            while let Ok(pos) = task::block_in_place(|| recv.recv()) {
-                debug!("io thread receive chunk pos {pos:?}");
-                level
-                    .chunk_saver
-                    .fetch_chunks(&level.level_folder, &[pos], t_send.clone())
-                    .await;
-                if let Some(Loaded(chunk)) = t_recv.recv().await {
-                    if send
-                        .send((pos, RecvChunk::IO(Chunk::Level(chunk))))
-                        .is_err()
-                    {
-                        log::info!("io thread stop");
-                        break;
-                    }
-                } else if send
-                    .send((
-                        pos,
-                        RecvChunk::IO(Chunk::Proto(StagedChunk::new(
-                            pos,
-                            gen_settings_from_dimension(&level.world_gen.dimension),
-                            level.world_gen.default_block,
-                            biome_mixer_seed,
-                        ))),
-                    ))
-                    .is_err()
-                {
-                    log::info!("io thread stop");
-                    break;
-                }
-            }
-        }
-
-        fn generation_work(
-            recv: Receiver<(ChunkPos, Cache, StagedChunkEnum)>,
-            send: Sender<(ChunkPos, RecvChunk)>,
-            level: Arc<Level>,
-        ) {
-            log::info!(
-                "generation thread start id: {:?} name: {}",
-                thread::current().id(),
-                thread::current().name().unwrap_or("unknown")
-            );
-
-            let settings = gen_settings_from_dimension(&level.world_gen.dimension);
-            while let Ok((pos, mut cache, stage)) = recv.recv() {
-                debug!("generation thread receive chunk pos {pos:?} to stage {stage:?}");
-                cache.advance(
-                    stage,
-                    level.block_registry.as_ref(),
-                    settings,
-                    &level.world_gen.random_config,
-                    &level.world_gen.terrain_cache,
-                    &level.world_gen.base_router,
-                    level.world_gen.dimension,
-                );
-                if send.send((pos, RecvChunk::Generation(cache))).is_err() {
-                    log::info!(
-                        "generation thread stop id: {:?} name: {}",
-                        thread::current().id(),
-                        thread::current().name().unwrap_or("unknown")
-                    );
-                    break;
-                }
-            }
-        }
-
-        fn drop_mark(&mut self, stage: StagedChunkEnum, pos: ChunkPos) {
-            let mark = self.task_mark.get_mut(&pos).unwrap();
-            debug_assert!((*mark >> (stage as u8) & 1) == 1);
-            *mark -= 1 << (stage as u8);
-            if *mark == 0 {
-                self.task_mark.remove(&pos);
-            }
-        }
-
-        fn receive_chunk(&mut self, pos: ChunkPos, data: RecvChunk) {
-            debug!("receive chunk pos {pos:?}");
-            match data {
-                RecvChunk::IO(chunk) => match chunk {
-                    Chunk::Level(data) => {
-                        self.loaded_chunks.insert(pos, data);
-                        self.task_mark.remove(&pos);
-                    }
-                    Chunk::Proto(data) => {
-                        self.proto_chunks.insert(pos, data);
-                        // i think we don't need to remove mark here
-                    }
-                },
-                RecvChunk::Generation(data) => {
-                    let mut dx = 0;
-                    let mut dy = 0;
-                    for chunk in data.chunks {
-                        let new_pos = ChunkPos::new(data.x + dx, data.y + dy);
-                        match chunk {
-                            Chunk::Level(chunk) => {
-                                if new_pos == pos {
-                                    // other chunk is borrowed by arc. don't need to return
-                                    self.loaded_chunks.insert(new_pos, chunk);
-                                    self.task_mark.remove(&new_pos);
-                                }
-                            }
-                            Chunk::Proto(chunk) => {
-                                self.proto_chunks.insert(new_pos, chunk);
-                            }
-                        }
-                        self.occupied.remove(&new_pos);
-                        dy += 1;
-                        if dy == data.size {
-                            dy = 0;
-                            dx += 1;
-                        }
-                    }
-                }
-            }
-            self.running_task_count -= 1;
-        }
-
-        fn dump_debug_info(&self, sx: i32, tx: i32, sy: i32, ty: i32) {
-            debug!("queue len {}", self.queue.len());
-            // debug!("queue {:?}", self.queue);
-            debug!("running tasks {}", self.running_task_count);
-            let mut s = String::new();
-            for x in sx..=tx {
-                for y in sy..=ty {
-                    s += Self::get_chunk_stage_id(
-                        &self.loaded_chunks,
-                        &self.proto_chunks,
-                        ChunkPos::new(x, y),
-                    )
-                    .to_string()
-                    .as_str();
-                    s += " ";
-                }
-                s += "\n";
-            }
-            debug!("chunk stage:\n{s}\n");
-
-            s.clear();
-            for x in sx..=tx {
-                for y in sy..=ty {
-                    s += self
-                        .last_level
-                        .get(&ChunkPos::new(x, y))
-                        .unwrap_or(&ChunkLoading::MAX_LEVEL)
-                        .to_string()
-                        .as_str();
-                    s += " ";
-                }
-                s += "\n";
-            }
-            debug!("last loading level:\n{s}\n");
-        }
-
-        fn work(mut self) {
-            log::info!(
-                "schedule thread start id: {:?} name: {}",
-                thread::current().id(),
-                thread::current().name().unwrap_or("unknown")
-            );
-            if let Some(new_level) = self.send_level.wait_and_get() {
-                debug!("receive new level");
-                self.resort_work(new_level);
-            }
-            let mut clock = Instant::now();
-            loop {
-                let mut nothing = true;
-                let mut len = self.queue.len();
-                let mut i = 0;
-
-                let now = Instant::now();
-                if now - clock > Duration::from_secs(5) {
-                    self.dump_debug_info(-20, 20, -20, 20);
-                    clock = now;
-                }
-                'outer: while i < len {
-                    while let Ok((pos, data)) = self.recv_chunk.try_recv() {
-                        self.receive_chunk(pos, data);
-                    }
-
-                    let (pos, _, stage) = self.queue[i];
-
-                    let level = *self
-                        .last_level
-                        .get(&pos)
-                        .unwrap_or(&ChunkLoading::MAX_LEVEL);
-                    if level == ChunkLoading::MAX_LEVEL {
-                        self.drop_mark(stage, pos);
-                        self.queue.remove(i);
-                        len -= 1;
-                        continue;
-                    }
-
-                    let highest_stage = StagedChunkEnum::level_to_stage(level);
-                    if (highest_stage as u8) < (stage as u8) {
-                        self.drop_mark(stage, pos);
-                        self.queue.remove(i);
-                        len -= 1;
-                        continue;
-                    }
-
-                    let current_stage =
-                        Self::get_chunk_stage_id(&self.loaded_chunks, &mut self.proto_chunks, pos);
-                    if current_stage >= (stage as u8) {
-                        self.drop_mark(stage, pos);
-                        self.queue.remove(i);
-                        len -= 1;
-                        continue;
-                    }
-
-                    if stage == Empty {
-                        nothing = false;
-                        self.running_task_count += 1;
-                        self.io.send(pos).expect("oi thread close unexpectedly");
-                        self.queue.remove(i);
-                        len -= 1;
-                        continue;
-                    }
-
-                    let radius = stage.get_radius();
-                    let depend = stage.get_dependencies();
-                    for dx in -radius..=radius {
-                        for dy in -radius..=radius {
-                            let new_pos = pos.add_raw(dx, dy);
-                            let dst = max(abs(dx), abs(dy)) as usize;
-                            if self.occupied.contains(&new_pos)
-                                || Self::get_chunk_stage_id(
-                                    &self.loaded_chunks,
-                                    &mut self.proto_chunks,
-                                    new_pos,
-                                ) < (depend[dst] as u8)
-                            {
-                                i += 1;
-                                continue 'outer;
-                            }
-                        }
-                    }
-                    let mut cache = Cache::new(pos.x - radius, pos.y - radius, (radius << 1) + 1);
-                    for dx in -radius..=radius {
-                        for dy in -radius..=radius {
-                            let new_pos = pos.add_raw(dx, dy);
-                            cache.chunks.push(
-                                Self::get_chunk(
-                                    &self.loaded_chunks,
-                                    &mut self.proto_chunks,
-                                    new_pos,
-                                )
-                                .expect("Chunk missing"),
-                            );
-                            self.occupied.insert(new_pos);
-                        }
-                    }
-                    self.running_task_count += 1;
-                    self.generate
-                        .send((pos, cache, stage))
-                        .expect("oi thread close unexpectedly");
-                    self.queue.remove(i);
-                    len -= 1;
-                }
-                if len == 0 {
-                    debug!("the queue is empty. thread sleep");
-                    while self.running_task_count > 0 {
-                        while let Ok((pos, data)) = self.recv_chunk.recv() {
-                            self.receive_chunk(pos, data);
-                        }
-                    }
-                    if let Some(new_level) = self.send_level.wait_and_get() {
-                        debug!("receive new level");
-                        self.resort_work(new_level);
-                    }
-                } else if let Some(new_level) = self.send_level.get() {
-                    debug!("receive new level");
-                    self.resort_work(new_level);
-                } else if nothing && self.running_task_count > 0 {
-                    debug!("nothing to do. thread sleep.");
-                    if let Ok((pos, data)) = self.recv_chunk.recv() {
-                        self.receive_chunk(pos, data);
-                    }
-                }
-            }
-        }
-    }
-}
 pub struct ChunkRequest {
     pub pos: Vector2<i32>,
     pub response: oneshot::Sender<(SyncChunk, bool)>, // bool = is_new
@@ -1166,8 +60,8 @@ pub struct ChunkRequest {
 /// For more details on world generation, refer to the `WorldGenerator` module.
 pub struct Level {
     pub seed: Seed,
-    block_registry: Arc<dyn BlockRegistryExt>,
-    level_folder: LevelFolder,
+    pub block_registry: Arc<dyn BlockRegistryExt>,
+    pub level_folder: LevelFolder,
 
     /// Counts the number of ticks that have been scheduled for this world
     schedule_tick_counts: AtomicU64,
@@ -1180,10 +74,10 @@ pub struct Level {
 
     chunk_watchers: Arc<DashMap<Vector2<i32>, usize>>,
 
-    chunk_saver: Arc<dyn FileIO<Data = SyncChunk>>,
+    pub chunk_saver: Arc<dyn FileIO<Data = SyncChunk>>,
     entity_saver: Arc<dyn FileIO<Data = SyncEntityChunk>>,
 
-    world_gen: Arc<VanillaGenerator>,
+    pub world_gen: Arc<VanillaGenerator>,
 
     /// Tracks tasks associated with this world instance
     tasks: TaskTracker,
@@ -1191,13 +85,12 @@ pub struct Level {
     pub shutdown_notifier: Notify,
     pub is_shutting_down: AtomicBool,
 
-    gen_request_tx: Sender<Vector2<i32>>,
-    pending_generations: Arc<DashMap<Vector2<i32>, Vec<oneshot::Sender<SyncChunk>>>>,
-
     gen_entity_request_tx: Sender<Vector2<i32>>,
     pending_entity_generations: Arc<DashMap<Vector2<i32>, Vec<oneshot::Sender<SyncEntityChunk>>>>,
 
-    pub sender: Mutex<Option<Arc<LevelChannel>>>,
+    pub level_channel: Arc<LevelChannel>,
+    pub thread_tracker: Mutex<Vec<std::thread::JoinHandle<()>>>,
+    pub chunk_listener: Arc<ChunkListener>,
 }
 
 pub struct TickData {
@@ -1237,7 +130,6 @@ impl Level {
         };
 
         // TODO: Load info correctly based on world format type
-
         let seed = Seed(seed as u64);
         let world_gen = get_world_gen(seed, dimension).into();
 
@@ -1257,11 +149,13 @@ impl Level {
                 }
             };
 
-        let (gen_request_tx, gen_request_rx) = crossbeam::channel::unbounded();
-        let pending_generations = Arc::new(DashMap::new());
-
         let (gen_entity_request_tx, gen_entity_request_rx) = crossbeam::channel::unbounded();
         let pending_entity_generations = Arc::new(DashMap::new());
+
+        let level_channel = Arc::new(LevelChannel::new());
+        let thread_tracker = Mutex::new(Vec::new());
+        let listener = Arc::new(ChunkListener::new());
+
         let level_ref = Arc::new(Self {
             seed,
             block_registry,
@@ -1272,61 +166,29 @@ impl Level {
             schedule_tick_counts: AtomicU64::new(0),
             loaded_chunks: Arc::new(DashMap::new()),
             loaded_entity_chunks: Arc::new(DashMap::new()),
-            chunk_loading: Mutex::new(ChunkLoading::new()),
+            chunk_loading: Mutex::new(ChunkLoading::new(level_channel.clone())),
             chunk_watchers: Arc::new(DashMap::new()),
             tasks: TaskTracker::new(),
             shutdown_notifier: Notify::new(),
             is_shutting_down: AtomicBool::new(false),
-            gen_request_tx,
-            pending_generations: pending_generations.clone(),
             gen_entity_request_tx,
             pending_entity_generations: pending_entity_generations.clone(),
-            sender: Mutex::new(None),
+            level_channel: level_channel.clone(),
+            thread_tracker,
+            chunk_listener: listener.clone(),
         });
 
-        let b = GenerationSchedule::new(&level_ref.tasks, 1, 12, level_ref.clone());
-        *level_ref.sender.lock().unwrap() = Some(b);
-
-        //TODO: Investigate optimal number of threads
         let num_threads = num_cpus::get().saturating_sub(1).max(1);
 
-        // Normal Chunks
-        for thread_id in 0..num_threads {
-            let level_clone = level_ref.clone();
-            let pending_clone = pending_generations.clone();
-            let rx = gen_request_rx.clone();
-
-            std::thread::spawn(move || {
-                while let Ok(pos) = rx.recv() {
-                    if level_clone.is_shutting_down.load(Ordering::Relaxed) {
-                        break;
-                    }
-
-                    log::debug!(
-                        "Generating chunk {pos:?}, worker thread {thread_id:?}, queue length {}",
-                        rx.len()
-                    );
-
-                    // Generate chunk
-                    let chunk = level_clone.world_gen.generate_chunk(
-                        &level_clone,
-                        level_clone.block_registry.as_ref(),
-                        &pos,
-                    );
-                    let arc_chunk = Arc::new(RwLock::new(chunk));
-
-                    // Insert into loaded chunks
-                    level_clone.loaded_chunks.insert(pos, arc_chunk.clone());
-
-                    // Fulfill all waiters
-                    if let Some(waiters) = pending_clone.remove(&pos) {
-                        for tx in waiters.1 {
-                            let _ = tx.send(arc_chunk.clone());
-                        }
-                    }
-                }
-            });
-        }
+        GenerationSchedule::create(
+            &level_ref.tasks,
+            4,
+            num_threads,
+            level_ref.clone(),
+            level_channel,
+            listener,
+            level_ref.thread_tracker.lock().unwrap().as_mut(),
+        );
 
         // Entity Chunks
         for thread_id in 0..num_threads {
@@ -1369,6 +231,7 @@ impl Level {
             .lock()
             .unwrap()
             .add_ticket(Vector2::<i32>::new(0, 0), 17);
+        // .add_ticket(Vector2::<i32>::new(0, 0), ChunkLoading::FULL_CHUNK_LEVEL);
         level_ref
     }
 
@@ -1407,6 +270,19 @@ impl Level {
 
         self.is_shutting_down.store(true, Ordering::Relaxed);
         self.shutdown_notifier.notify_waiters();
+        self.level_channel.notify();
+
+        let mut lock = self.thread_tracker.lock().unwrap();
+        log::info!("Wait {} jobs stop", lock.len());
+        while !lock.is_empty() {
+            let i = lock.pop().unwrap();
+            log::info!(
+                "Waiting Thread {:?} {} stop",
+                i.thread().id(),
+                i.thread().name().unwrap_or("unknown")
+            );
+        }
+        log::info!("All Thread stop");
 
         self.tasks.close();
         log::debug!("Awaiting level tasks");
@@ -1417,16 +293,16 @@ impl Level {
         self.chunk_saver.block_and_await_ongoing_tasks().await;
 
         // save all chunks currently in memory
-        let chunks_to_write = self
-            .loaded_chunks
-            .iter()
-            .map(|chunk| (*chunk.key(), chunk.value().clone()))
-            .collect::<Vec<_>>();
-        self.loaded_chunks.clear();
+        // let chunks_to_write = self
+        //     .loaded_chunks
+        //     .iter()
+        //     .map(|chunk| (*chunk.key(), chunk.value().clone()))
+        //     .collect::<Vec<_>>();
+        // self.loaded_chunks.clear();
 
         // TODO: I think the chunk_saver should be at the server level
-        self.chunk_saver.clear_watched_chunks().await;
-        self.write_chunks(chunks_to_write).await;
+        // self.chunk_saver.clear_watched_chunks().await;
+        // self.write_chunks(chunks_to_write).await;
 
         log::debug!("Done awaiting level entity tasks");
 
@@ -1483,9 +359,9 @@ impl Level {
             }
         }
 
-        self.chunk_saver
-            .watch_chunks(&self.level_folder, chunks)
-            .await;
+        // self.chunk_saver
+        //     .watch_chunks(&self.level_folder, chunks)
+        //     .await;
         self.entity_saver
             .watch_chunks(&self.level_folder, chunks)
             .await;
@@ -1521,10 +397,6 @@ impl Level {
                 }
             }
         }
-
-        self.chunk_saver
-            .unwatch_chunks(&self.level_folder, chunks)
-            .await;
         self.entity_saver
             .unwatch_chunks(&self.level_folder, chunks)
             .await;
@@ -1538,6 +410,7 @@ impl Level {
     }
 
     pub async fn clean_chunks(self: &Arc<Self>, chunks: &[Vector2<i32>]) {
+        panic!();
         // Care needs to be take here because of interweaving case:
         // 1) Remove chunk from cache
         // 2) Another player wants same chunk
@@ -1709,8 +582,8 @@ impl Level {
 
     pub fn clean_memory(&self) {
         self.chunk_watchers.retain(|_, watcher| !watcher.is_zero());
-        self.loaded_chunks
-            .retain(|at, _| self.chunk_watchers.get(at).is_some());
+        // self.loaded_chunks
+        //     .retain(|at, _| self.chunk_watchers.get(at).is_some());
         self.loaded_entity_chunks
             .retain(|at, _| self.chunk_watchers.get(at).is_some());
 
@@ -1734,34 +607,28 @@ impl Level {
     pub async fn get_chunk(self: &Arc<Self>, pos: Vector2<i32>) -> SyncChunk {
         // Already loaded?
         if let Some(chunk) = self.loaded_chunks.get(&pos) {
-            return chunk.clone();
+            chunk.clone()
         } else {
-            panic!("not chunk found at {pos:?}");
-        }
+            log::info!("Missing Chunk {pos:?}. Fetching.");
+            let recv = self.chunk_listener.add_single_chunk_listener(pos);
 
-        // Try to load from disk
-        match self.load_single_chunk(pos).await {
-            Ok((chunk, _)) => {
-                self.loaded_chunks.insert(pos, chunk.clone());
-                chunk
+            {
+                let mut lock = self.chunk_loading.lock().unwrap();
+                lock.add_ticket(pos, ChunkLoading::FULL_CHUNK_LEVEL);
+                lock.send_change();
             }
-            Err(_) => {
-                // Need to generate
-                let (tx, rx) = oneshot::channel();
 
-                // Deduplication
-                match self.pending_generations.entry(pos) {
-                    dashmap::mapref::entry::Entry::Occupied(mut entry) => {
-                        entry.get_mut().push(tx);
-                    }
-                    dashmap::mapref::entry::Entry::Vacant(entry) => {
-                        entry.insert(vec![tx]);
-                        let _ = self.gen_request_tx.send(pos);
-                    }
-                }
+            let ret = if let Some(chunk) = self.loaded_chunks.get(&pos) {
+                // try again here. otherwise deadlock
+                chunk.clone()
+            } else {
+                recv.await.unwrap()
+            };
 
-                rx.await.expect("Generation worker dropped")
-            }
+            let mut lock = self.chunk_loading.lock().unwrap();
+            lock = self.chunk_loading.lock().unwrap();
+            lock.remove_ticket(pos, ChunkLoading::FULL_CHUNK_LEVEL);
+            ret
         }
     }
 
@@ -1773,6 +640,7 @@ impl Level {
         self: &Arc<Self>,
         chunks: Vec<Vector2<i32>>,
     ) -> UnboundedReceiver<(SyncChunk, bool)> {
+        panic!("");
         let (sender, receiver) = mpsc::unbounded_channel();
         let level = self.clone();
 
@@ -1820,23 +688,23 @@ impl Level {
                                 // Need to generate — but don't block here
                                 let level_clone = level.clone();
 
-                                tokio::spawn(async move {
-                                    let (tx, rx) = oneshot::channel();
-
-                                    match level_clone.pending_generations.entry(pos) {
-                                        dashmap::mapref::entry::Entry::Occupied(mut entry) => {
-                                            entry.get_mut().push(tx);
-                                        }
-                                        dashmap::mapref::entry::Entry::Vacant(entry) => {
-                                            entry.insert(vec![tx]);
-                                            let _ = level_clone.gen_request_tx.send(pos);
-                                        }
-                                    }
-
-                                    if let Ok(chunk) = rx.await {
-                                        let _ = sender_clone.send((chunk, true));
-                                    }
-                                });
+                                // tokio::spawn(async move {
+                                //     let (tx, rx) = oneshot::channel();
+                                //
+                                //     match level_clone.pending_generations.entry(pos) {
+                                //         dashmap::mapref::entry::Entry::Occupied(mut entry) => {
+                                //             entry.get_mut().push(tx);
+                                //         }
+                                //         dashmap::mapref::entry::Entry::Vacant(entry) => {
+                                //             entry.insert(vec![tx]);
+                                //             let _ = level_clone.gen_request_tx.send(pos);
+                                //         }
+                                //     }
+                                //
+                                //     if let Ok(chunk) = rx.await {
+                                //         let _ = sender_clone.send((chunk, true));
+                                //     }
+                                // });
                             }
                         }
                     }
