+++ pumpkin-world/src/lib.rs
@@ -6,6 +6,7 @@ use pumpkin_util::math::vector2::Vector2;
 pub mod biome;
 pub mod block;
 pub mod chunk;
+pub mod chunk_system;
 pub mod cylindrical_chunk_iterator;
 pub mod data;
 pub mod dimension;
